pp/sr: fine grained result package
Prerequisite for extracting logic from schema registry.
