+++ src/v/pandaproxy/schema_registry/json.cc
@@ -21,6 +21,7 @@
 #include "pandaproxy/schema_registry/error.h"
 #include "pandaproxy/schema_registry/errors.h"
 #include "pandaproxy/schema_registry/json_schema/dialect.h"
+#include "pandaproxy/schema_registry/result.h"
 #include "pandaproxy/schema_registry/sharded_store.h"
 #include "pandaproxy/schema_registry/types.h"
